+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Tags no longer require a type prefix. If there is no prefix, it refers to the
   property on the object itself.
+- Name overrides now use PipeWire match rules.
 
 ### Fixed